Keep an empty temp/ folder for intermediate outputs
Following a convention to add an empty .gitkeep file to initialize an otherwise empty directory.  Without force adding .gitkeep to the repo, the temp/ directory would not be initialized due to .gitignore.  Note that the filename .gitkeep is just a convention but does not otherwise have any special behavior in git.