Write first draft of volgistics matching workflow (UNTESTED)
Adapting the matching notebook and other experiments into the production Python workflow.  The code has not yet been tested due to some Python module issues, and there are a few details left to work out on the exact matching rules, how the code is invoked, and related details before it can be merged into master.