Outline the pseudocode for #81
Still need to modularize the code, refactor into match_data.py,
delete the older (now unused) functions, and test it.  But, the
overall outline will now implement the full matching rules in an
extensible way that allows PAWS to specify the priority order for
different data sources based on reliability.