Successfully run matching script on new master_df rows
Fixed a bug in the left/right variables for the table join step.
Also cleaned up the output format for the next step in the flow_script
pipeline. We should discuss a few questions as a group (and implement
master_df table in postgres) before this feature will be fully
complete, but it's in substantially better shape now than it was.